GG-22316 Fixed JCache TCK

+++ modules/core/src/test/java/org/apache/ignite/internal/processors/security/cache/EntryProcessorPermissionCheckTest.java
@@ -78,7 +78,7 @@ public class EntryProcessorPermissionCheckTest extends AbstractCacheOperationPer
 
         c.accept(CACHE_NAME, entry);
 
-        assertThat(verifierNode.cache(CACHE_NAME).get(entry.getKey()), is(entry.getValue()));
+        assertThat((Integer)(verifierNode.cache(CACHE_NAME).get(entry.getKey())), is(entry.getValue()));
     }
 
     /** */
